+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `EveOnline::ESI::CharacterWalletTransactions`
 * Rename `EveOnline::ESI::Models::WalletJournalEntry` to `EveOnline::ESI::Models::WalletJournal`
 * Add `EveOnline::ESI::Models::Killmail`
+* Rename `EveOnline::ESI::Models::WalletJournal#id` to `#wallet_journal_id` and update `#as_json`
 
 ## v0.31.0